requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
charset-normalizer==3.4.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pdf2image==1.17.0
pillow==11.0.0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
requests==2.32.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
tornado==6.1
tzlocal==5.2
urllib3==2.2.3
